+++ esp-idf-svc/CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - OTA - Implements a new type `EspFirmwareInfoLoad` that has a reduced memory consumption (#531)
 - Added set_promiscuous function in EthDriver (#246)
 - Added set_promiscuous, is_promiscuous functions in WifiDriver (#246)
+- Blocking and buffered StdIo (#541) - i.e. easy reading/input from `std::io::stdin`
 
 ### Fixed
 - The alloc::Vec version stomps the scan state from Done to Idle. (#459)